moving stuff to correct directory
