+++ client/containers/Pub/PubDocument/PubBottom/pubBottom.scss
@@ -1,4 +1,4 @@
-@import '../pub.scss';
+@import '../../pub.scss';
 
 .pub-bottom-component {
     background: #F6F4F4;
